pyproject.toml
[tool.poetry]
name = "votr-flask"
version = "0.1.0"
description = ""
authors = ["Andrew Chong <i@andrewhchong.com>",
            "Sebastien Allen <allenseb@oregonstate.edu"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.0.2"
Flask-SQLAlchemy = "^2.5.1"
SQLAlchemy = "^1.4.31"
Flask-Migrate = "^3.1.0"
gunicorn = "^20.1.0"
Flask-WTF = "^1.0.0"
WTForms = "^3.0.1"
Jinja2 = "^3.0.3"
python-dotenv = "^0.19.2"
PyMySQL = "^1.0.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.1.0"
flake8 = "^4.0.1"
Flask-DebugToolbar = "^0.11.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
